elapsed_time argument in Update methods is now a float data type

+++ ttcards/src/CardsMenuState.cpp
@@ -189,7 +189,7 @@ void CardsMenu::onMouseWheel ( bool up, bool down )
   }
 }
 
-void CardsMenu::Update ( uint32_t elapsed_time )
+void CardsMenu::Update ( float elapsed_time )
 {
   this->updateCursor();
